Make file paths relative to the game directory
This makes a fairly significant change to the way that scripts will be able to access files. The first time you run FiveTool, you will be asked to browse for your game folder. Any time a script accesses a file, it will be relative to this folder.

This provides a couple of key advantages.

First of all, it provides a means of sandboxing script file access. One of my goals with FiveTool is to ensure that scripts are always safe to run. By requiring that all paths be relative to the game directory, scripts won't be able to access anything outside of it or gain information about the user's system. (I will be adding some mechanisms in the future that allow access to other files with the user's explicit approval.)

Second, it simplifies things for users. There's no longer any need to extract FiveTool to your game directory and clutter up everything. You can put FiveTool anywhere you want and scripts will always access the correct files. It also makes typing easier when you're using the interactive console because you won't have to remember where your game files are at.

In order to make this system as simple as possible from a development perspective, I created a static FileAccessUtil.ResolvePath() method in C#. Call it in any code that accesses files and you'll be able to ensure that paths are handled securely.