File dialog functions
These functions allow scripts to ask the user to choose a file anywhere on their system. They will grant scripts a "path token", which is a unique string that allows them to refer to the selected file as long as they have the token. I hope this will strike a good balance between safety and ease of use.

For basic use, you can use the ChooseFile() and ChooseNewFile() functions, but the internal Dialog.OpenFile() function is also available for more advanced users.